review: Вынес реалезацию стека слоев из класса приложения. Теперь приложение при создании получает слои в виде ILayerCollection.
