+++ template/core/application.py
@@ -6,23 +6,23 @@ from core.application_startup_options import ApplicationStartupOptions
 
 from core.timestep import Timestep
 from core.renderer.window import Window
-from core.renderer.window_viewport import WindowViewSpecification
+from core.renderer.window import WindowViewSpecification
 from core.renderer.window_toolkit import WindowToolkit
 from core.events.event_dispatcher import EventDispatcher
 from core.layers.layer import Layer
-from core.layers.layer_stack import LayerStack
+from core.layers.ilayer_collection import ILayerCollection
 
 
 class Application:
 
-    def __init__(self, options: ApplicationStartupOptions) -> NoReturn:
+    def __init__(self, options: ApplicationStartupOptions, layers: ILayerCollection) -> NoReturn:
         self.__options = options
 
         self.__prepare_platform()
 
         self.__clock = pygame.time.Clock()
 
-        self.__layer_stack = LayerStack()
+        self.__layers = layers
         self.__running = True
         self.__title = self.__options.title
 
@@ -38,10 +38,7 @@ class Application:
     @property
     def window_toolkit(self) -> WindowToolkit:
         return self.__window_toolkit
-
-    def add_layer(self, layer: Layer) -> NoReturn:
-        self.__layer_stack.add_layer(layer)
-
+        
     def run(self) -> NoReturn:
         self.__begin_session()
         while self.__running:
@@ -61,8 +58,8 @@ class Application:
         self.__window.prepare(pygame.Color(18, 18, 18))
 
         time_step = Timestep(self.__clock.get_time())
-        self.__layer_stack.on_(lambda layer: layer.on_update(time_step))
-        self.__layer_stack.on_(lambda layer: layer.on_render(render_context))
+        self.__layers.on_(lambda layer: layer.on_update(time_step))
+        self.__layers.on_(lambda layer: layer.on_render(render_context))
 
         self.__window.render()
 
@@ -79,7 +76,7 @@ class Application:
         dispatcher.dispatch(pygame.QUIT, self.__on_close)
         dispatcher.dispatch_key_down(pygame.K_ESCAPE, self.__on_close)
 
-        self.__layer_stack.on_(lambda layer: layer.on_event(event))
+        self.__layers.on_(lambda layer: layer.on_event(event))
 
     def __prepare_platform(self) -> NoReturn:
         pygame.init()
@@ -89,8 +86,8 @@ class Application:
         pygame.quit()
 
     def __begin_session(self) -> NoReturn:
-        self.__layer_stack.on_(lambda layer: layer.on_attach())
+        self.__layers.on_(lambda layer: layer.on_attach())
 
     def __finish_session(self) -> NoReturn:
         self.__shutdown_platform()
-        self.__layer_stack.on_(lambda layer: layer.on_detach())
+        self.__layers.on_(lambda layer: layer.on_detach())
