+++ template/core/layers/layer_stack.py
@@ -1,19 +1,22 @@
 from typing import Callable, NoReturn
 
 from core.layers.layer import Layer
+from core.layers.ilayer_collection import ILayerCollection
 
 
-class LayerStack:
+class LayerStack(ILayerCollection):
 
     def __init__(self) -> NoReturn:
         self.__layers = list()
 
-    @property
-    def layers(self) -> list[Layer]:
-        return self.__layers
-
     def add_layer(self, layer: Layer) -> NoReturn:
         self.__layers.append(layer)
+
+    def remove_layer(self, layer: Layer) -> NoReturn:
+        if not layer in self.__layers:
+            raise ValueError(f"Layer Stack does not contain layer: {layer}. Yoc cannot remove it.")
+
+        self.__layers.pop(self.__layers.index(layer))
 
     def on_(self, action: Callable) -> NoReturn:
         for layer in self.__layers:
