+++ template/core/renderer/window.py
@@ -4,7 +4,25 @@ from typing import Callable, NoReturn
 
 from core.renderer.window_viewport import WindowViewport
 from core.renderer.window_viewport import WindowViewFormat
-from core.renderer.window_viewport import WindowViewSpecification
+
+
+class WindowViewSpecification:
+
+    def __init__(self, resolution: tuple[int, int], smoothed: bool) -> NoReturn:
+        self.__smoothed = smoothed
+        self.__resolution = resolution
+
+    @property
+    def is_smoothed(self) -> bool:
+        return self.__smoothed
+
+    @property
+    def resolution(self) -> tuple[int, int]:
+        return self.__resolution
+
+    @property
+    def resolution_as_vector(self) -> pygame.Vector2:
+        return pygame.Vector2(self.__resolution)
 
 
 class Window:
